+++ upchatpy/api.py
@@ -23,14 +23,14 @@ SOFTWARE.
 """
 
 import logging
-from datetime import datetime, timedelta
+from datetime import UTC, datetime, timedelta
 from typing import AsyncGenerator, Optional
 from urllib.parse import urlencode
-from uuid import UUID
 
 import aiohttp
 
 from .exceptions import AuthenticationError, HTTPError, ResourceNotFoundError
+from .responses.auth import AuthResponse
 from .responses.orders import Order, OrderItem, OrderResponse, OrdersResponse
 from .responses.products import ProductResponse, ProductsResponse
 from .responses.users import UsersResponse
@@ -44,7 +44,7 @@ log = logging.getLogger("upgrade.chat")
 class Client:
     BASE_URL = "https://api.upgrade.chat"
 
-    def __init__(self, client_id: str, client_secret: str):
+    def __init__(self, client_id: str, client_secret: str, auth: Optional[AuthResponse] = None):
         """
         Initializes the Client with the provided client ID and client secret.
 
@@ -54,14 +54,17 @@ class Client:
         """
         self.client_id = client_id
         self.client_secret = client_secret
-        self.access_token: Optional[str] = None
+        self.auth = auth
 
-    async def get_auth(self) -> None:
+    async def get_auth(self) -> AuthResponse:
         """
         Authenticates the client and retrieves the access token from Upgrade.Chat.
 
         Raises:
             AuthenticationError: if authentication fails.
+
+        Returns:
+            AuthResponse: The authentication response.
         """
         url = f"{self.BASE_URL}/oauth/token"
         data = {
@@ -70,11 +73,11 @@ class Client:
             "grant_type": "client_credentials",
         }
         async with aiohttp.ClientSession() as session:
-            async with session.post(url=url, json=data) as res:
-                if res.status != 200:
-                    raise AuthenticationError(res.status, "Failed to authenticate with Upgrade.Chat API")
-                results = await res.json()
-                self.access_token = results["access_token"]
+            async with session.post(url=url, json=data) as response:
+                if response.status != 200:
+                    raise AuthenticationError(response.status, "Failed to authenticate with Upgrade.Chat API")
+                self.auth = AuthResponse.model_validate(await response.json())
+                return self.auth
 
     async def _request(self, method: str, endpoint: str, data: dict = None) -> dict:
         """
@@ -92,16 +95,20 @@ class Client:
         Returns:
             dict: The JSON response as a dictionary.
         """
-        if self.access_token is None:
+        if self.auth is None:
+            log.debug("No auth token, fetching")
+            await self.get_auth()
+        elif self.auth.access_token_expired:
+            log.debug("Access token expired, refreshing")
             await self.get_auth()
 
-        headers = {"Authorization": f"Bearer {self.access_token}"}
+        headers = {"Authorization": f"Bearer {self.auth.access_token}"}
         url = f"{self.BASE_URL}{endpoint}"
 
         try:
             async with aiohttp.ClientSession(headers=headers) as session:
                 async with session.request(method, url, data=data) as response:
-                    log.debug(f"{method} ({response.status}): {url}")
+                    log.debug("%s, (%s): %s", method, response.status, url)
 
                     if response.status == 404:
                         error_details = await response.json()
@@ -127,7 +134,7 @@ class Client:
         Fetches orders with pagination support.
 
         Args:
-            limit (int, optional): The maximum number of orders to retrieve per page. Defaults to 100.
+            limit (int, optional): The maximum number of orders to retrieve per page. Defaults to 100 (max is 100).
             offset (int, optional): The offset to start retrieving orders from. Defaults to 0.
             user_discord_id (Optional[str], optional): Filter orders for a specific Discord user ID. Defaults to None.
             order_type (Optional[str], optional): Filter orders by type. Defaults to None.
@@ -139,6 +146,8 @@ class Client:
         Yields:
             Iterator[AsyncGenerator[OrdersResponse, None]]: OrdersResponse object
         """
+        if not 1 <= limit <= 100:
+            raise ValueError("Limit must be between 1 and 100")
         offset = offset or 0
         while True:
             res = await self.get_orders(limit, offset, user_discord_id, order_type, coupon)
@@ -160,7 +169,7 @@ class Client:
         Fetches a list of orders from the Upgrade.Chat API.
 
         Args:
-            limit (int, optional): The maximum number of orders to retrieve per page. Defaults to 100.
+            limit (int, optional): The maximum number of orders to retrieve per page. Defaults to 100 (max is 100).
             offset (int, optional): The offset to start retrieving orders from. Defaults to 0.
             user_discord_id (Optional[str], optional): Filter orders for a specific Discord user ID. Defaults to None.
             order_type (Optional[str], optional): Filter orders by type. Defaults to None.
@@ -169,6 +178,8 @@ class Client:
         Returns:
             OrdersResponse: An OrdersResponse object containing the fetched orders.
         """
+        if not 1 <= limit <= 100:
+            raise ValueError("Limit must be between 1 and 100")
         query_params = {
             "limit": limit,
             "offset": offset,
@@ -202,7 +213,7 @@ class Client:
         Fetches products with pagination support.
 
         Args:
-            limit (int, optional): The number of products to fetch per request. Defaults to 100.
+            limit (int, optional): The number of products to fetch per request. Defaults to 100 (max is 100).
             offset (int, optional): The offset from where to start fetching products. Defaults to 0.
             product_type (Optional[str], optional): Optional product type to filter products. Defaults to None.
 
@@ -212,6 +223,8 @@ class Client:
         Yields:
             Iterator[AsyncGenerator[ProductsResponse, None]]: ProductsResponse object.
         """
+        if not 1 <= limit <= 100:
+            raise ValueError("Limit must be between 1 and 100")
         offset = offset or 0
         while True:
             res = await self.get_products(limit, offset, product_type)
@@ -228,13 +241,15 @@ class Client:
         Asynchronously fetches products with pagination support.
 
         Args:
-            limit (int, optional): The number of products to fetch per request. Defaults to 100.
+            limit (int, optional): The number of products to fetch per request. Defaults to 100 (max is 100).
             offset (int, optional): The offset from where to start fetching products. Defaults to 0.
             product_type (Optional[str], optional): Optional product type to filter products. Defaults to None.
 
         Returns:
             ProductsResponse: A ProductsResponse object containing the fetched products.
         """
+        if not 1 <= limit <= 100:
+            raise ValueError("Limit must be between 1 and 100")
         query_params = {
             "limit": limit,
             "offset": offset,
@@ -264,12 +279,14 @@ class Client:
         """Fetches a list of users from the Upgrade.Chat API.
 
         Args:
-            limit (int, optional): The maximum number of users to retrieve. Defaults to 100.
+            limit (int, optional): The maximum number of users to retrieve. Defaults to 100 (max is 100).
             offset (int, optional): The offset to start retrieving users from. Defaults to 0.
 
         Returns:
             UsersResponse: A UsersResponse object containing the fetched users.
         """
+        if not 1 <= limit <= 100:
+            raise ValueError("Limit must be between 1 and 100")
         query_params = {"limit": limit, "offset": offset}
         query_string = urlencode(query_params)
         endpoint = f"/v1/users?{query_string}"
@@ -282,12 +299,14 @@ class Client:
         Fetches a list of webhooks with pagination support
 
         Args:
-            limit (int): The maximum number of webhooks to retrieve.
+            limit (int): The maximum number of webhooks to retrieve, defaults to 100 (max is 100).
             offset (int): The offset to start retrieving webhooks from, this will increment by the limit each iteration.
 
         Returns:
             AsyncGenerator[WebhooksResponse, None]: An async generator yielding WebhooksResponse objects.
         """
+        if not 1 <= limit <= 100:
+            raise ValueError("Limit must be between 1 and 100")
         offset = offset or 0
         while True:
             res = await self.get_webhooks(limit, offset)
@@ -302,12 +321,14 @@ class Client:
         Fetches a list of webhooks.
 
         Args:
-            limit (int): The maximum number of webhooks to retrieve.
+            limit (int): The maximum number of webhooks to retrieve, defaults to 100 (max is 100).
             offset (int): The offset to start retrieving webhooks from.
 
         Returns:
             WebhooksResponse: A WebhooksResponse object containing the fetched webhooks.
         """
+        if not 1 <= limit <= 100:
+            raise ValueError("Limit must be between 1 and 100")
         query_params = {"limit": limit, "offset": offset}
         query_string = urlencode(query_params)
         endpoint = f"/v1/webhooks?{query_string}"
@@ -315,12 +336,12 @@ class Client:
         response = await self._request("GET", endpoint)
         return WebhooksResponse.model_validate(response)
 
-    async def get_webhook(self, webhook_id: UUID) -> WebhookResponse:
+    async def get_webhook(self, webhook_id: str) -> WebhookResponse:
         """
         Fetches a single webhook by ID.
 
         Args:
-            webhook_id (UUID): The ID of the webhook to retrieve.
+            webhook_id (str): The ID of the webhook to retrieve.
 
         Returns:
             WebhookResponse: A WebhookResponse object containing the webhook details.
@@ -335,12 +356,14 @@ class Client:
         Fetches a list of webhooks events with pagination support
 
         Args:
-            limit (int): The maximum number of webhooks to retrieve.
+            limit (int): The maximum number of webhooks to retrieve, defaults to 100 (max is 100).
             offset (int): The offset to start retrieving webhooks from, this will increment by the limit each iteration.
 
         Returns:
             AsyncGenerator[WebhookEventsResponse, None]: An async generator yielding WebhookEventsResponse objects.
         """
+        if not 1 <= limit <= 100:
+            raise ValueError("Limit must be between 1 and 100")
         offset = offset or 0
         while True:
             res = await self.get_webhook_events(limit, offset)
@@ -355,12 +378,14 @@ class Client:
         Fetches a list of webhook events.
 
         Args:
-            limit (int): The maximum number of webhook events to retrieve.
+            limit (int): The maximum number of webhook events to retrieve, defaults to 100 (max is 100).
             offset (int): The offset to start retrieving webhook events from.
 
         Returns:
             WebhookEventsResponse: A WebhookEventsResponse object containing the fetched webhook events.
         """
+        if not 1 <= limit <= 100:
+            raise ValueError("Limit must be between 1 and 100")
         query_params = {"limit": limit, "offset": offset}
         query_string = urlencode(query_params)
         endpoint = f"/v1/webhook-events?{query_string}"
@@ -368,12 +393,12 @@ class Client:
         response = await self._request("GET", endpoint)
         return WebhookEventsResponse.model_validate(response)
 
-    async def get_webhook_event(self, event_id: UUID) -> WebhookEventResponse:
+    async def get_webhook_event(self, event_id: str) -> WebhookEventResponse:
         """
         Fetches a single webhook event by ID.
 
         Args:
-            event_id (UUID): The ID of the webhook event to retrieve.
+            event_id (str): The ID of the webhook event to retrieve.
 
         Returns:
             WebhookEventResponse: A WebhookEventResponse object containing the webhook event details.
@@ -381,12 +406,12 @@ class Client:
         response = await self._request("GET", f"/v1/webhook-events/{event_id}")
         return WebhookEventResponse.model_validate(response)
 
-    async def validate_webhook_event(self, event_id: UUID) -> WebhookValidResponse:
+    async def validate_webhook_event(self, event_id: str) -> WebhookValidResponse:
         """
         Validates a webhook event by ID.
 
         Args:
-            event_id (UUID): The ID of the webhook event to validate.
+            event_id (str): The ID of the webhook event to validate.
 
         Returns:
             WebhookValidResponse: A WebhookValidResponse object indicating if the event is valid.
@@ -396,7 +421,7 @@ class Client:
 
     async def user_is_subscribed(
         self,
-        product_uuid: UUID,
+        product_uuid: str,
         user_discord_id: str,
         cancelled_with_time_left: bool = True,
         ignore_not_found: bool = True,
@@ -405,7 +430,7 @@ class Client:
         Checks if a user is currently subscribed to a product.
 
         Args:
-            product_uuid (UUID): The UUID of the product to check.
+            product_uuid (str): The UUID of the product to check.
             user_discord_id (str): The Discord ID of the user to check.
             cancelled_with_time_left (bool, optional): If true, will consider a cancelled subscription with time left as still subscribed. Defaults to True.
             ignore_not_found (bool, optional): If false, will raise an exception if the user does not exist. Defaults to True.
@@ -414,6 +439,7 @@ class Client:
             bool: True if the user is subscribed to the product, False otherwise.
         """
         user_orders: list[Order] = []
+        product_uuid = product_uuid.lower()
 
         try:
             async for orders in self.aget_orders(user_discord_id=user_discord_id, order_type="UPGRADE"):
@@ -430,11 +456,11 @@ class Client:
         except ResourceNotFoundError:
             if not ignore_not_found:
                 raise
-            log.debug(f"User {user_discord_id} does not exist in Upgrade.Chat")
+            log.debug("User %s does not exist in Upgrade.Chat", user_discord_id)
             return False
 
         if not user_orders:
-            log.debug(f"User {user_discord_id} has no orders for product {product_uuid}")
+            log.debug("User %s has no orders for product %s", user_discord_id, product_uuid)
             return False
 
         # Sort orders by purchased_at date with the most recent first
@@ -442,33 +468,33 @@ class Client:
 
         most_recent_order: Order = user_orders[0]
         if not most_recent_order.order_items or not most_recent_order.purchased_at:
-            log.debug(f"User {user_discord_id} has no order items or purchase date for product {product_uuid}")
+            log.debug("User %s has no order items or purchase date for product %s", user_discord_id, product_uuid)
             return False
 
         order_item: OrderItem = most_recent_order.order_items[0]
 
         # Check if first order is still active
         if not most_recent_order.cancelled_at:
-            log.debug(f"User {user_discord_id} is subscribed to product {product_uuid}")
+            log.debug("User %s is subscribed to product %s", user_discord_id, product_uuid)
             return True
 
         # If first order is cancelled, check if there is still time left on the subscription based on its interval
         elif most_recent_order.cancelled_at and cancelled_with_time_left:
             purchased_at = most_recent_order.purchased_at
-            interval = order_item.interval
-            interval_count = order_item.interval_count
+            interval = order_item.interval.value
+            interval_count = order_item.interval_count or 1
             # Based on the purchase date, determine if they have left on their subscription
             if interval == "day":
-                return purchased_at + timedelta(days=interval_count) > datetime.utcnow()
+                return purchased_at + timedelta(days=interval_count) > datetime.now(UTC)
             elif interval == "week":
-                return purchased_at + timedelta(weeks=interval_count) > datetime.utcnow()
+                return purchased_at + timedelta(weeks=interval_count) > datetime.now(UTC)
             elif interval == "month":
-                return purchased_at + timedelta(months=interval_count) > datetime.utcnow()
+                return purchased_at + timedelta(days=interval_count * 30) > datetime.now(UTC)
             elif interval == "year":
-                return purchased_at + timedelta(years=interval_count) > datetime.utcnow()
+                return purchased_at + timedelta(days=interval_count * 365) > datetime.now(UTC)
             else:
-                log.error(f"Unknown interval type for order {most_recent_order.uuid}, interval: {interval}")
+                log.error("Unknown interval type for order %s, interval: %s", most_recent_order.uuid, interval)
                 return False
 
-        log.debug(f"User {user_discord_id} is not subscribed to product {product_uuid}")
+        log.debug("User %s is not subscribed to product %s", user_discord_id, product_uuid)
         return False